refactor(BA-4926): extract CustomIntrospectionRule to shared module
Deduplicate CustomIntrospectionRule that was identically defined in both
the REST admin handler and the GQL validation extension by moving it to
manager/api/graphql_rules.py.
