+++ src/ai/backend/manager/api/gql/extensions/validation.py
@@ -1,26 +1,13 @@
 from __future__ import annotations
 
 from collections.abc import Iterator
-from typing import Any
 
-from graphql import GraphQLError, ValidationRule
+from graphql import ValidationRule
 from strawberry.extensions.base_extension import SchemaExtension
 from strawberry.extensions.query_depth_limiter import create_validator as create_depth_validator
 
 from ai.backend.manager.api.gql.types import StrawberryGQLContext
-
-
-class CustomIntrospectionRule(ValidationRule):
-    """Blocks introspection queries (fields starting with ``__``) except ``__typename``."""
-
-    def enter_field(self, node: Any, *_args: Any) -> None:
-        field_name = node.name.value
-        if field_name.startswith("__"):
-            if field_name == "__typename":
-                return
-            self.report_error(
-                GraphQLError(f"Cannot query '{field_name}': introspection is disabled.", node)
-            )
+from ai.backend.manager.api.graphql_rules import CustomIntrospectionRule
 
 
 class GQLValidationExtension(SchemaExtension):
